+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.0.0
 
+* Added `.factory` to define factory functions
 * Rewrite everything with ES6
 * Targeting nodejs `^4.0`